refactor: move About window to Windows folder
