+++ Windows/AboutWindow.xaml
@@ -1,4 +1,4 @@
-<Window x:Class="DamnSimpleFileManager.AboutWindow"
+<Window x:Class="DamnSimpleFileManager.Windows.AboutWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         Title="About" Height="200" Width="300"
